refactor: 将列表页和详情页，提取通用组件
GenericListPage.tsx：用于列表页（page.tsx）
GenericPostPage.tsx：用于详情页（[slug]/page.tsx）
