+++ app/blog/page.tsx
@@ -1,33 +1,10 @@
-import { PostList } from 'app/components/PostList'
-import { getPostFiles } from 'app/utils/utils'
-import { notFound } from 'next/navigation'
+import GenericListPage, {
+  generateListMetadata,
+} from 'app/components/GenericListPage'
 import { folder } from './config'
 
-import { siteTitle } from 'app/config'
-import { metadataBase } from '../components/metadataBase'
-
-const folderTitle = folder.replace(/\b(\s\w|^\w)/g, function (txt) {
-  return txt.toUpperCase()
-})
-
-export const metadata = {
-  ...metadataBase,
-  title: folderTitle + ' - ' + siteTitle,
-}
+export const metadata = generateListMetadata(folder)
 
 export default function Page() {
-  const list = getPostFiles([folder])
-
-  if (list.length === 0) {
-    notFound()
-  }
-
-  return (
-    <section>
-      <h1 className="font-semibold text-2xl mb-8 tracking-tighter">
-        {folderTitle}
-      </h1>
-      <PostList list={list} />
-    </section>
-  )
+  return <GenericListPage folder={folder} />
 }
