+++ app/changelog/config.ts
@@ -1,0 +1,4 @@
+/**
+ * 当前文件夹名
+ */
+export const folder = 'changelog'
